# swagger-config.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'BookStore API'
  description: 'An REST API of book store. Allows you to CRUD books, authors, genres, publishers. Also you can create an account and make orders'
  contact:
    name: Ravil
    email: rvl.c137@gmail.com
servers:
  - url: "http://localhost:8080"
paths:
...
  /genres:
    description: Genre resource. Get, post genres
    get:
      tags:
        - genres
      description: Get all genres from the database
      parameters:
      - name: page
        required: false
        in: query
        description: The page which you want to get
        schema:
          type: integer
          example: 0
      - name: size
        required: false
        in: query
        description: The amount of elements in one page
        schema:
          type: integer
          example: 10
      responses:
        200:
          description: Successfully returned an array of genres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/genre'
    post:
      tags:
        - genres
      description: Save a genre to the database
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/genre'
      responses:
        200:
          description: Successfully created a genre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/genre'
  /genres/{id}:
    get:
      tags:
        - genres
      description: Get specific genre by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: Successfully retrieved genre from database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/genre'
        400:
          description: Invalid id format!
        403:
          description: Forbidden!
        404:
          description: Genre not found. Invalid id supplied!
    put:
      tags:
        - genres
      description: Update genre
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/genre'
      responses:
        200:
          description: Successfully updated a genre
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/genre'
        400:
          description: Invalid id format!
        403:
          description: Forbidden!
        404:
          description: Genre not found. Invalid id supplied!
    delete:
      tags:
        - genres
      description: Delete genre by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 2
      responses:
        200:
          description: Successfully deleted a genre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/genre'
        400:
          description: Invalid id format!
        403:
          description: Forbidden!
        404:
          description: Genre not found. Invalid id supplied!
...
  /orders:
    get:
      description: Get all orders of currently logged in user
      tags:
        - orders
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            example: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            example: 10
      responses:
        200:
          description: Successfully returned an array of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/order'
        403:
          description: Forbidden. User not authenticated!
    post:
      tags:
        - orders
      description: Create a new order
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/saveOrder'
      responses:
        200:
          description: Successfully created an order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order'
        403:
          description: Forbidden. User not authenticated!
  /orders/{id}:
    get:
      tags:
        - orders
      description: Get specific order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 2
      responses:
        200:
          description: Successfully retrieved order from the database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order'
        400:
          description: Invalid id format!
        403:
          description: Forbidden. Only users with role ADMIN can get information about other orders
        404:
          description: Order not found. Invalid id supplied!
    put:
      tags:
        - orders
      description: Method for admins to change order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/saveOrder'
      responses:
        200:
          description: Successfully updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order'
        400:
          description: Invalid id format!
        403:
          description: Forbidden! Only users with role ADMIN can change status of an order
        404:
          description: User not found. Invalid id supplied!
    delete:
      tags:
        - orders
      description: Delete a order with specific id. Order record does not erases from the database, rather order status changes to 'canceled'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 2
      responses:
        200:
          description: Successfully deleted order
        400:
          description: Invalid id format!
        403:
          description: Forbidden! Only users with role ADMIN can delete orders
        404:
          description: Order not found. Invalid id supplied!
  /orders/findAll:
    get:
      description: Get all orders from the database
      tags:
        - orders
      parameters:
      - name: page
        required: false
        in: query
        description: The page which you want to get
        schema:
          type: integer
          example: 0
      - name: size
        required: false
        in: query
        description: The amount of elements in one page
        schema:
          type: integer
          example: 10
      responses:
        200:
          description: Successfully returned an array of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/order'
        403:
          description: Forbidden. Only users with role admin can retrieve information about all orders
  /orders/putOwn/{id}:
    description: Resourse for users to change their own orders
    put:
      tags:
        - orders
      description: Method for users to change their own order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/saveOrder'
      responses:
        200:
          description: Successfully updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order'
        400:
          description: Invalid id format!
        403:
          description: Forbidden! User not authorized.
        404:
          description: User not found. Invalid id supplied!
  /orders/changeStatus:
    description: Method for admins to change order status
    put:
      tags:
        - orders
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: integer
            example: 2
        - name: newStatus
          in: query
          required: true
          schema:
            type: string
            example: approved
      responses:
        200:
          description: Successfully changed status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order'
        400:
          description: Invalid id format!
        403:
          description: Forbidden! Only users with role admin can change status of the order
        404:
          description: User not found. Invalid id supplied!
components:
  schemas:
    book:
      type: object
      properties:
        id:
          type: integer
          example: 1
        price:
          type: number
          format: double
          example: 4000.0
        authorList:
          type: array
          items:
            type: string
            example: ['Fyodor Dostoevsky', 'Howard Lovecraft']
        publisher:
          type: string
          example: 'Azbuka'
        title:
          type: string
          example: 'Clean Code'
        numberOfPages:
          type: integer
          example: 300
        releaseYear:
          type: integer
          example: 1992
        genreList:
          type: array
          items:
            type: string
            example: ['fiction', 'children literature ']
        bookQuantity:
          type: integer
          example: 10
...
    genre:
      type: object
      properties:
        genreId:
          type: integer
          example: 1
        genreName:
          type: string
          example: 'thriller'
    order:
      type: object
      properties:
        orderID:
          type: integer
          example: 1
        userId:
          type: integer
          example: 1
        orderStatus:
          type: string
          example: 'created'
        orderTime:
          type: string
          format: date-time
          example: '2021-01-30T08:30:00Z'
        totalPrice:
          type: number
          example: 9982.5
        orderedBooks:
          type: array
          items:
            $ref: '#/components/schemas/book'
        bookAmount:
          type: array
          items:
            type: integer
            example: [1, 2]
    saveOrder:
      type: object
      properties:
        orderedBooks:
          type: array
          items:
            type: integer
            example: [1, 2]
        bookAmount:
          type: array
          items:
            type: integer
            example: [1, 2]
...